backend/requirements.txt
# Django e Core
Django==5.2.8
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2

# REST API
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# Banco de Dados
mysqlclient==2.2.4

# CORS e Segurança
django-cors-headers==4.5.0

# Produção
gunicorn==21.2.0

# Dev (opcional)
python-decouple==3.8  # Para gerenciar variáveis de ambiente